pyproject.toml
[tool.poetry]
name = "authentication"
version = "0.1.0"
description = ""
authors = ["Bruno Borelli <brunoborellipro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.4"
tzdata = "^2024.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.2"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . | ruff check . --diff'
format = 'ruff check . --fix | ruff format .'
run = 'fastapi dev authentication/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=authentication -vv'
post_test = 'coverage html'
